+++ src/diff_robot_control/launch/gazebo.launch.py
@@ -56,13 +56,6 @@ def generate_launch_description():
         ),
         
         Node(
-            package='joint_state_publisher',
-            executable='joint_state_publisher',
-            name='joint_state_publisher',
-            parameters=[{'use_sim_time': use_sim_time}]
-        ),
-        
-        Node(
             package='controller_manager',
             executable='spawner',
             arguments=['joint_state_broadcaster', '--controller-manager', '/controller_manager'],